requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
smtplib